pyproject.toml
[project]
name = "riseapp"
version = "0.1.0"
description = "Data lineage detection for XYZ Bank - external/internal docs, chunking, embeddings, FAISS, ADK agent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk",
    "google-genai",
    "pydantic>=2.0",
    "structlog",
    "pyyaml",
    "python-dotenv",
    "numpy",
    "faiss-cpu",
    "pymupdf",
    "sentence-transformers",
    "pytest",
    "pytest-asyncio",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio"]

[project.scripts]
riseapp = "main:main"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["test"]
asyncio_mode = "auto"
